+++ CHANGELOG.md
@@ -19,6 +19,7 @@ the code was deployed.
 
 - BraveAlerterConfigurator to reject the RESET state (CU-860r8k57h).
 - `num_button_presses` column to be renamed to `number_of_alerts` to match BraveSensor (CU-86791yyvg).
+- Button status changes Sentry log to become Twilio text messages (CU-860raw0d7).
 
 ## [13.5.0] - 2023-01-04